+++ R/check_design.R
@@ -81,7 +81,6 @@ check_design <- R6::R6Class(
         "Check task aliases cannot have the same name as any of the available packages" = !any(df$alias %in% available.packages(repos = repos)[, "Package"]),
         "Custom package aliases cannot be duplicates of check aliases" = !any(uulist(drlapply(df$custom, `[[`, "alias")) %in% df$alias)
       )
-
       if (!restore) unlink(output, recursive = TRUE, force = TRUE)
 
       self$input <- df
@@ -100,11 +99,17 @@ check_design <- R6::R6Class(
     active_processes = function() {
       private$active
     },
-
+    
+    #' @description
+    #' Get Failed Tasks list
+    failed_tasks = function() {
+      private$failed
+    },
+    
     #' @description
     #' Terminate Design Processes
     #'
-    #' Immedaitely termiantes all the active processes.
+    #' Immediately terminates all the active processes.
     terminate = function() {
       invisible(lapply(private$active, function(process) process$kill()))
     },
@@ -178,7 +183,8 @@ check_design <- R6::R6Class(
     repos = getOption("repos"),
     # active processes
     active = list(),
-
+    # failed tasks
+    failed = list(),
     # Methods
 
     push_process = function(task, x) {
@@ -186,7 +192,9 @@ check_design <- R6::R6Class(
       name <- task_graph_task_name(self$graph, task)
       task_graph_package_status(self$graph, task) <- STATUS$`in progress`
       x$set_finalizer(function(process) {
-        # TODO: Implement warning if the process failed before finalizing
+        if (process$get_r_exit_status() != 0) {
+          private$failed[[name]] <- task
+        }
         private$pop_process(name)
         task_graph_package_status(self$graph, task) <- STATUS$done
       })